Remove Instrument constant and moved instrument image fetch logic
