+++ src/components/search/instrument/instrument-search-item.tsx
@@ -1,4 +1,5 @@
-import { Instrument } from "@/constants/instruments"
+import { getInstrumentImage } from "@/helpers/cloudinary.helper"
+import { Instrument } from "@/types/database-types"
 import Image from "next/image"
 
 interface InstrumentDropdownItemProps {
@@ -14,8 +15,10 @@ const InstrumentSearchItem = ({ textNode, instrument, onClick }: InstrumentDropd
       className="flex items-center p-4 hover:bg-moon-300 hover:text-white"
     >
       <Image
-        src={instrument.icon}
-        alt={instrument.instrumentName.toString()}
+        src={getInstrumentImage(instrument.image_source)}
+        width={64}
+        height={64}
+        alt={instrument.instrument_name.toString()}
         className={"mr-4 h-6 w-6 hover:text-white"}
       />
       {textNode}
